Refactor IAM policies for size
Add a new policy file for app security and move inspector and waf(v2)
into it.  Accordingly; make a related terminator file.

Condense ec2:Describe* policies down to a wildcard.  This has the effect
of adding numerous new policies but I believe they're ok, having reviewed
the current full list of ec2:Describe* actions.

Also adds DMS endpoint support, which has been blocked in #139 waiting on this
policy rework.
