+++ aws/policy/application-security.yaml
@@ -1,0 +1,117 @@
+Version: '2012-10-17'
+Statement:
+
+  - Sid: AllowRegionalRestrictedResourceActionsWhichIncurFees
+    Effect: Allow
+    Action:
+      - wafv2:ListRuleGroups
+      - wafv2:ListWebACLs
+      - wafv2:AssociateWebACL
+      - wafv2:DeleteRuleGroup
+      - wafv2:CreateRuleGroup
+      - wafv2:PutFirewallManagerRuleGroups
+      - wafv2:GetWebACLForResource
+      - wafv2:GetLoggingConfiguration
+      - wafv2:DeleteWebACL
+      - wafv2:GetRateBasedStatementManagedKeys
+      - wafv2:ListLoggingConfigurations
+      - wafv2:GetIPSet
+      - wafv2:CreateWebACL
+      - wafv2:ListIPSets
+      - wafv2:GetWebACL
+      - wafv2:GetRuleGroup
+      - wafv2:CreateIPSet
+      - wafv2:ListAvailableManagedRuleGroups
+      - wafv2:DeleteIPSet
+      - wafv2:DescribeManagedRuleGroup
+      - wafv2:CheckCapacity
+      - wafv2:ListResourcesForWebACL
+      - wafv2:DeleteLoggingConfiguration
+      - wafv2:PutLoggingConfiguration
+      - wafv2:DisassociateWebACL
+      - wafv2:UpdateWebACL
+      - wafv2:UpdateRuleGroup
+      - wafv2:DeleteFirewallManagerRuleGroups
+      - wafv2:DisassociateFirewallManager
+      - wafv2:UpdateIPSet
+      - wafv2:TagResource
+    Resource:
+      - 'arn:aws:wafv2:{{ aws_region }}:{{ aws_account_id }}:*'
+
+  - Sid: AllowRegionalUnrestrictedResourceActionsWhichIncurNoFees
+    Effect: Allow
+    Action:
+      - inspector:ListAssessmentTargets
+      - inspector:CreateResourceGroup
+      - inspector:CreateAssessmentTarget
+      - inspector:DescribeAssessmentTargets
+      - inspector:DescribeResourceGroups
+      - inspector:UpdateAssessmentTarget
+      - inspector:DeleteAssessmentTarget
+      - inspector:ListAssessmentTemplates
+      - inspector:ListRulesPackages
+      - inspector:DescribeRulesPackages
+      - inspector:CreateAssessmentTemplate
+      - inspector:DeleteAssessmentTemplate
+      - inspector:SetTagsForResource
+      - inspector:DescribeAssessmentTemplates
+      - inspector:ListTagsForResource
+      - inspector:ListEventSubscriptions
+      - waf:CreateByteMatchSet
+      - waf:CreateGeoMatchSet
+      - waf:CreateIPSet
+      - waf:CreateRateBasedRule
+      - waf:CreateRegexMatchSet
+      - waf:CreateRegexPatternSet
+      - waf:CreateRule
+      - waf:CreateRuleGroup
+      - waf:CreateSizeConstraintSet
+      - waf:CreateSqlInjectionMatchSet
+      - waf:CreateWebACL
+      - waf:CreateXssMatchSet
+      - waf:DeleteByteMatchSet
+      - waf:DeleteGeoMatchSet
+      - waf:DeleteIPSet
+      - waf:DeleteRateBasedRule
+      - waf:DeleteRegexMatchSet
+      - waf:DeleteRegexPatternSet
+      - waf:DeleteRule
+      - waf:DeleteRuleGroup
+      - waf:DeleteSizeConstraintSet
+      - waf:DeleteSqlInjectionMatchSet
+      - waf:DeleteWebACL
+      - waf:DeleteXssMatchSet
+      - waf:GetByteMatchSet
+      - waf:GetChangeToken
+      - waf:GetChangeTokenStatus
+      - waf:GetGeoMatchSet
+      - waf:GetIPSet
+      - waf:GetRateBasedRule
+      - waf:GetRateBasedRuleManagedKeys
+      - waf:GetRegexMatchSet
+      - waf:GetRegexPatternSet
+      - waf:GetRule
+      - waf:GetRuleGroup
+      - waf:GetSizeConstraintSet
+      - waf:GetSqlInjectionMatchSet
+      - waf:GetWebACL
+      - waf:GetXssMatchSet
+      - waf:List*
+      - waf:TagResource
+      - waf:UntagResource
+      - waf:UpdateByteMatchSet
+      - waf:UpdateGeoMatchSet
+      - waf:UpdateIPSet
+      - waf:UpdateRateBasedRule
+      - waf:UpdateRegexMatchSet
+      - waf:UpdateRegexPatternSet
+      - waf:UpdateRule
+      - waf:UpdateSizeConstraintSet
+      - waf:UpdateSqlInjectionMatchSet
+      - waf:UpdateWebACL
+      - waf:UpdateXssMatchSet
+    Resource: "*"
+    Condition:
+      StringEquals:
+        aws:RequestedRegion:
+          - '{{ aws_region }}'
