+++ aws/policy/compute.yaml
@@ -70,20 +70,7 @@ Statement:
       - ec2:DeleteTags
       - ec2:DeleteTransitGateway
       - ec2:DeregisterImage
-      - ec2:DescribeAccountAttributes
-      - ec2:DescribeIamInstanceProfileAssociations
-      - ec2:DescribeImageAttribute
-      - ec2:DescribeImages
-      - ec2:DescribeInstanceAttribute
-      - ec2:DescribeInstances
-      - ec2:DescribeInstanceStatus
-      - ec2:DescribeKeyPairs
-      - ec2:DescribeLaunchTemplates
-      - ec2:DescribeLaunchTemplateVersions
-      - ec2:DescribeRegions
-      - ec2:DescribeSnapshots
-      - ec2:DescribeTags
-      - ec2:DescribeVolumes
+      - ec2:Describe*
       - ec2:DetachVolume
       - ec2:DisassociateIamInstanceProfile
       - ec2:GetLaunchTemplateData
@@ -118,9 +105,6 @@ Statement:
       - autoscaling:DescribeInstanceRefreshes
       - autoscaling:DescribeLaunchConfigurations
       - autoscaling:DescribePolicies
-      - ec2:DescribeAvailabilityZones
-      - ec2:DescribeSpotPriceHistory
-      - ec2:DescribeTransitGateways
       - elasticloadbalancing:DeleteRule
       - elasticloadbalancing:DescribeListeners
       - elasticloadbalancing:DeleteListener
