+++ aws/policy/networking.yaml
@@ -62,26 +62,6 @@ Statement:
       - ec2:DeleteVpcPeeringConnection
       - ec2:DeleteVpnConnection
       - ec2:DeleteVpnGateway
-      - ec2:DescribeAddresses
-      - ec2:DescribeCustomerGateways
-      - ec2:DescribeDhcpOptions
-      - ec2:DescribeEgressOnlyInternetGateways
-      - ec2:DescribeInternetGateways
-      - ec2:DescribeNatGateways
-      - ec2:DescribeNetworkAcls
-      - ec2:DescribeNetworkInterfaces
-      - ec2:DescribeRouteTables
-      - ec2:DescribeSecurityGroups
-      - ec2:DescribeSubnets
-      - ec2:DescribeVpcAttribute
-      - ec2:DescribeVpcClassicLink
-      - ec2:DescribeVpcClassicLinkDnsSupport
-      - ec2:DescribeVpcEndpoints
-      - ec2:DescribeVpcEndpointServices
-      - ec2:DescribeVpcPeeringConnections
-      - ec2:DescribeVpcs
-      - ec2:DescribeVpnConnections
-      - ec2:DescribeVpnGateways
       - ec2:DetachInternetGateway
       - ec2:DetachNetworkInterface
       - ec2:DetachVpnGateway
